3 Managed passing props from Parent FormCarPhoto to child FC FormSearch and FC CarPhoto
