+++ src/components/CarPhoto/CarPhoto.tsx
@@ -2,7 +2,6 @@ import React, { FC } from 'react';
 import 'the-new-css-reset';
 import * as card from './CarPhoto.styled';
 import * as plate from './CardPhotoPlate.styled';
-import { StringDecoder } from 'string_decoder';
 
 interface IdPhotoProps {
    _id: number | string,
@@ -16,7 +15,6 @@ interface SizePhotoProps {
 
 interface CarPhotoProps extends SizePhotoProps, IdPhotoProps {
    _digits?: string,
-   _vendor?: string,
    _model?: string,
    _model_year?: number | string,
    _registered_at?: string,
@@ -25,21 +23,21 @@ interface CarPhotoProps extends SizePhotoProps, IdPhotoProps {
 
 const getUrlPhoto = (url_photo: string | undefined) => {
    
-   if (url_photo === ''){
+   if (url_photo === 'soon'){
       return './img/car_come_soon.jpg';
    }
    else if (url_photo === ('error404')){
       return './img/error404.jpg';
    }
 
-   return url_photo;
+   return './img/car_icon.jpg';
 }
 
 const CarPhoto: FC<CarPhotoProps> = (props) => {
 
    return (
       <card.CarPhotoWrapper _width={props._width_photo} _height={props._height_photo} _border='main'>
-        <card.CarPhotoImg src={require(`${getUrlPhoto(props._url_photo)}`)}/>
+        <card.CarPhotoImg src={props._url_photo != ''? props._url_photo : require(`${getUrlPhoto(props._url_photo)}`)}/>
         <card.CarPhotoInner _direction='column'>
            <card.ContentTop>
               <plate.Number>
@@ -60,13 +58,13 @@ const CarPhoto: FC<CarPhotoProps> = (props) => {
                  <span>{props._registered_at || 'dd.mm.yyyy'}</span>
               </plate.DateRegistr>
            </card.ContentTop>
-           <p style={{color: 'red'}}>Test value&nbsp;{props._value}</p>
+           {/* <p style={{color: 'red'}}>Test value&nbsp;{props._value}</p> */}
            <card.ContentBottom _direction='column' _justify='end' _align='start'>
               <plate.DataMark>
                  <span>{props._model_year || 'x-yyyy'}</span>
               </plate.DataMark>
               <plate.DataMark>
-                 <span>{props._vendor || 'x-vendor'}&nbsp;{props._model || 'x-model'}</span>
+                 <span>{props._model || 'x-model'}</span>
               </plate.DataMark>
            </card.ContentBottom>
         </card.CarPhotoInner>
@@ -83,7 +81,6 @@ CarPhoto.defaultProps  = {
    _width_photo: 500,
    _height_photo: 375,
    _digits: 'AA DDDD AA',
-   _vendor: 'x-vendor',
    _model: 'x-model',
    _model_year: 'x-yyyy',
    _registered_at: 'dd.mm.yyyy',
