+++ src/components/FormCarPhoto/FormCarPhoto.styled.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
 
 export const FormCarPhotoWrapper = styled.div`
-    background-color: aqua;
+    background-color: rgb(236, 254, 220);
 `;
